backend/ai/pyproject.toml
[tool.poetry]
name = "near-ai"
version = "0.1.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
aiohttp = "^3.9.5"
pydantic = "^2.8.2"
nltk = "^3.8.1"
langchain = "^0.2.7"
python-socketio = "^5.11.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
